Moved device setup outside of knapsack solver to avoid repetition
Now the solver expects a dict of lists of the devices available, which is created by the setup_devices function.
This dict of lists only contains the relevant device indices, their hashrates, and their power consumptions; the corresponding energy usage is still calculated inside the solver function, because it depends on the timeframe.
The type checks have been adjusted respectively.

The solver also outputs a "solution" dictionary with the relevant outputs.

NOTE: from now on, we will assume that each device inside computers.json will have a "count" of 1.
It makes no sense to have multiple counts of a device, if the output script expects to address them individually anyway.
Future updates will remove the "count" property.